Moved managers to their namespace
