+++ Registry.Web.Test/ChunkedUploadManagerTest.cs
@@ -16,6 +16,7 @@ using Registry.Web.Data.Models;
 using Registry.Web.Models;
 using Registry.Web.Models.Configuration;
 using Registry.Web.Services.Adapters;
+using Registry.Web.Services.Managers;
 using Registry.Web.Services.Ports;
 
 namespace Registry.Web.Test
